client: Make cert and key PathLike
This changes `MAISClient.cert` and `.key` to accept an `os.PathLike` instead of
a `pathlib.Path`.  This allows clients to provide a `str`.

Also, a note is added that we will not recognize paths that are relate to
things like `~`; if clients have such a path, they should be resolved first.