docs/workgroup: Lots of content, and a restructuring
The modules under stanford.mais.workgroup are the largest part of this package,
by far.  As I was writing the documentation, I came to realize that the
'narrative text mixed in with class documentation' structure was not working
here.  So, I took inspiration from Python-Requests documentation, and wrote
separate narrative sections for each major part of workgroups:

* The API client (WorkgroupClient)

* The Workgroup Class, including basic operations and properties

* Workgroup Membership

* Privgroup

Module documentation is at the end, providing the docstrings for each class in
each module.

Coming soon, the other sections of the documentation will be changed to a
similar structure.