workgroup: Raise IndexError on early search wildcard
When you perform a "Search for workgroups by name" search, you are able to
include wildcards in your search.  There is an undocumented requirement: You
must have at least four characters before your first wildcard.  This minimum
can be a problem if you want to fetch a list of all workgroups in a short stem.
For example, `hr:*` will fail.

When this requirement is violated, the API returns a 400 code with a special
message.  This commit changes the code which handles the 400 response code:
Instead of immediately raising an exception, we first check for the location of
any wildcard: If the wildcard is too early in the search string, we now raise an
IndexError.

Why wait for the 400 response?  Because, MaIS has an internal ticket (WG-1186)
to examine changing the minimum.  If they change the minimum, I'd like the
change to be visible to users without needing to upgrade.

Besides the code change (which includes the docstring), this commit also adds a
testcase (using a newly-mocked response), and updates the Workgroup
documentation.

Updates #36.  This doesn't actually close it, though.