workgroup.tests: Mock privgroup for a private workgroup
When a client attempts to get the privgroup of a private workgroup, and the
client is not an administrator of the workgroup, the call will return an empty
privgroup, with a message about the workgroup being private.  Per #31, this is
intentional behavior that will not be changed.

This commit adds a mock for that API response.