workgroup: get_privgroup catches loss of access after fetch
`Workgroup.get_privgroup()` already included a check of `.can_see_membership`
at the start of the method: If that check fails, a PermissionError is
immediately raised to the caller.  This can happen if the workgroup is private,
and we are not an administrator.

But, what happens if we *were* an administrator at the time we fetched the
workgroup, but then we lost that access?  In that case, the API call would
succeed, but we would get an empty privgroup, along with the message "This is a
private workgroup!".

This is a Time of Check to Time of Use (TOCTOU) error.

This commit fixes that: After parsing the privgroup, if the privgroup contains
zero members and administrators, check for the special message.  If we got the
special message, then raise a PermissionError.  Otherwise, return the empty
privgroup.

This commit also updates the existing `test_get_private` test to catch this
possibility.

Closes: #31