client: Set a custom user-agent
Modify our requests custom session with a custom User-Agent header.
The user-agent will look something like this:

    Python/3.4.5 stanford-mais/1.2.3 requests/2.3.4

The Python verison is pulled from the `platform` module; the other two package
versions are pulled from `importlib.metadata`: That pacakge stabalized in
Python 3.10; since we require at least Python 3.11, it's safe to use, and means
we don't need to rely on pkg_resources.

We shouldn't expect pacakge versions to change out from under us, so the
user-agent is a static string, computed at import time.

If the client provides their own custom Session, then it'll be on the client to
set this header.  I update the `MAISClient` documentation to note this (and to
also note that the `Accept` header must also be set).

Finally, a test case is added to check the User-Agnet header (along with the
Accept header, which wasn't previously being tested).

Closes #46