workgroup: last_update ignores membership changes
The Workgroup's last_update datestamp only resets when a workgroup's property
is changes.  It does *not* reset on membership changes.  So, our documentation
is wrong.

This commit changes the `Workgroup.last_update` docstring to make it explicit
about what will cause this datestamp to reset.  This commit also changes the
narrative documenation, though the mention is not as explicit there.

Closes #45