workflow: Do pip-install in setup-python
The new version of the setup-python action includes the ability to run
`pip install …`.  I take advantage of this by having it run `pip install
--upgrade pip`.

There is a downside that this happens before the pip cache is restored,
but that shouldn't be a big deal, as the setup-python action generally
gets us the latest pip from the start.